Add fetching all event, events by Id and delete event queries in the events.ts file
